Guillaume : petites modifs + début import/export

+++ src/main/java/com/webapp/site/DefaultCategoryService.java
@@ -8,6 +8,7 @@ import javax.inject.Inject;
 import org.springframework.stereotype.Service;
 
 import com.webapp.site.entities.Category;
+import com.webapp.site.entities.User;
 import com.webapp.site.repositories.CategoryRepository;
 
 @Service
@@ -34,18 +35,21 @@ public class DefaultCategoryService implements CategoryService {
 	}
 
 	@Override
-	public void delete(long id) {
-		categorieRepository.deleteById(id);
-	}
-	
-	@Override
-	public Category getCategory(String name){
-		return categorieRepository.getOneByName(name);
+	public void delete(long id, String username) {
+		Category category = getCategory(id);
+		if(category.getUser().getUsername().equals(username)) {
+			categorieRepository.deleteById(id);
+		}
 	}
 	
 	@Override
 	public List<Category> getCategoriesByUsername(String username){
 		return categorieRepository.findByUser_username(username);
+	}
+	
+	@Override
+	public Category getCategoryByNameAndUsername(String name, String username) {
+		return this.categorieRepository.getOneByNameAndUser_username(name, username);
 	}
 
 }
